requirements.txt
django
djangorestframework
djongo
django-cors-headers
pymongo==3.12.3